requirements.txt
# Web
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==22.0.0

# Core image BG removal
rembg==2.0.56
numpy==1.23.5
Pillow==10.4.0
onnxruntime==1.19.2
opencv-python-headless==4.12.0.88
scikit-image==0.21.0
pymatting==1.1.14
imageio==2.35.1
tifffile==2023.7.10
PyWavelets==1.4.1
tqdm==4.67.1
pooch==1.8.2
jsonschema==4.23.0

# numba stack required by pymatting/scikit-image on py38
llvmlite==0.39.1
numba==0.56.4
